requirements.txt
lightgbm>=4.0
scikit-learn>=1.3
openai>=1.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
python-dotenv>=1.0
httpx>=0.27
rich>=13.0
